cleanup utils folder bazel dependency
