+++ src/ray/gcs/redis_context.h
@@ -22,6 +22,7 @@
 #include "ray/common/asio/instrumented_io_context.h"
 #include "ray/common/status.h"
 #include "ray/gcs/redis_async_context.h"
+#include "ray/util/exponential_backoff.h"
 #include "ray/util/logging.h"
 #include "ray/util/util.h"
 #include "src/ray/protobuf/gcs.pb.h"
