+++ src/ray/rpc/client_call.h
@@ -26,6 +26,7 @@
 #include "ray/common/grpc_util.h"
 #include "ray/common/id.h"
 #include "ray/common/status.h"
+#include "ray/util/thread_utils.h"
 #include "ray/util/util.h"
 
 namespace ray {
