+++ src/ray/util/exponential_backoff.cc
@@ -14,13 +14,11 @@
 
 #include "ray/util/exponential_backoff.h"
 
-#include <math.h>
-
-#include "ray/util/logging.h"
+#include <cmath>
 
 namespace ray {
 
-uint64_t ExponentialBackoff::GetBackoffMs(uint64_t attempt,
+uint64_t ExponentialBackOff::GetBackoffMs(uint64_t attempt,
                                           uint64_t base_ms,
                                           uint64_t max_backoff_ms) {
   uint64_t delay = static_cast<uint64_t>(pow(2, attempt));
