+++ src/ray/util/exponential_backoff.h
@@ -14,15 +14,33 @@
 
 #pragma once
 
-#include <gtest/gtest_prod.h>
-#include <stdint.h>
+#include <cstdint>
+#include <limits>
+
+#include "ray/util/logging.h"
 
 namespace ray {
 
 /// Provides the exponential backoff algorithm that is typically used
 /// for throttling.
-class ExponentialBackoff {
+class ExponentialBackOff {
  public:
+  /// Construct an exponential back off counter.
+  ///
+  /// \param[in] initial_value The start value for this counter
+  /// \param[in] multiplier The multiplier for this counter.
+  /// \param[in] max_value The maximum value for this counter. By default it's
+  ///    infinite double.
+  ExponentialBackOff(uint64_t initial_value,
+                     double multiplier,
+                     uint64_t max_value = std::numeric_limits<uint64_t>::max())
+      : curr_value_(initial_value),
+        initial_value_(initial_value),
+        max_value_(max_value),
+        multiplier_(multiplier) {
+    RAY_CHECK(multiplier > 0.0) << "Multiplier must be greater than 0";
+  }
+
   /// Computes the backoff delay using the exponential backoff algorithm,
   /// using the formula
   ///
@@ -33,11 +51,28 @@ class ExponentialBackoff {
   /// @return the delay in ms based on the formula
   static uint64_t GetBackoffMs(uint64_t attempt,
                                uint64_t base_ms,
-                               uint64_t max_backoff_ms = kMaxBackoffMs);
+                               uint64_t max_backoff_ms = kDefaultMaxBackoffMs);
+
+  uint64_t Next() {
+    auto ret = curr_value_;
+    curr_value_ = curr_value_ * multiplier_;
+    curr_value_ = std::min(curr_value_, max_value_);
+    return ret;
+  }
+
+  uint64_t Current() { return curr_value_; }
+
+  void Reset() { curr_value_ = initial_value_; }
 
  private:
+ private:
+  uint64_t curr_value_;
+  uint64_t initial_value_;
+  uint64_t max_value_;
+  double multiplier_;
+
   // The default cap on the backoff delay.
-  static constexpr uint64_t kMaxBackoffMs = 1 * 60 * 1000;
+  static constexpr uint64_t kDefaultMaxBackoffMs = 1 * 60 * 1000;
 };
 
 }  // namespace ray
